+++ opal/mca/pmix/pmix3x/pmix3x.c
@@ -1191,9 +1191,7 @@ int pmix3x_value_unload(opal_value_t *kv,
             /* handle the various types */
             if (PMIX_INFO == v->data.darray->type) {
                 pmix_info_t *iptr = (pmix_info_t*)v->data.darray->array;
-                if (NULL != iptr[n].key) {
-                    ival->key = strdup(iptr[n].key);
-                }
+                ival->key = strdup(iptr[n].key);
                 rc = pmix3x_value_unload(ival, &iptr[n].value);
                 if (OPAL_SUCCESS != rc) {
                     OPAL_LIST_RELEASE(lt);
